Working on data presentation.
This is first go at adding some syntax highlighting.
The tests are currently failing because the expected display data is
changing while fields are being added.

Probably need to consider using 2html and comparing against that for
testing, this would compare both the data and the syntax highlighting.  Also
it's easy to see what the buffer contents should be when opening in a browser.